Move StringOrFilePath from core to config
